# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "bluebell项目接口文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8084",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/author/articles/export": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "批量导出多篇文章为Markdown格式",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "导出"
                ],
                "summary": "批量导出文章",
                "parameters": [
                    {
                        "description": "导出请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExportArticlesBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "批量导出数据",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseExportBatch"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/author/articles/{id}/export": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "导出单篇文章为Markdown格式",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "导出"
                ],
                "summary": "导出单篇文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "导出数据",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseExportArticle"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/comments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取指定文章的评论列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "获取评论列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "article_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "评论列表",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseCommentList"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建新的评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "创建评论",
                "parameters": [
                    {
                        "description": "评论参数",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ParamCreateComment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建的评论",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseComment"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/comments/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新评论内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "更新评论",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "评论ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "评论内容",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除指定评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "删除评论",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "评论ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.ResCode": {
            "type": "integer",
            "format": "int64",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                1005,
                1006,
                1007,
                1008,
                1009,
                1010,
                1011,
                1012,
                1013,
                1014
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeInvalidParam",
                "CodeUserExist",
                "CodeUserNotExist",
                "CodeInvalidPassword",
                "CodeServerBusy",
                "CodeNeedLogin",
                "CodeInvalidToken",
                "CodeCategoryNotExist",
                "CodeCategoryExist",
                "CodeCategoryNameInvalid",
                "CodeNoPermissionCategory",
                "CodeArticleAlreadyInCategory",
                "CodeNoPermission",
                "CodeArticleNotExist"
            ]
        },
...
        "controller._ResponseComment": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "data": {},
                "msg": {}
            }
        },
        "controller._ResponseCommentList": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "list": {},
                        "page": {
                            "type": "integer"
                        },
                        "pages": {
                            "type": "integer"
                        },
                        "size": {
                            "type": "integer"
                        },
                        "total": {
                            "type": "integer"
                        }
                    }
                },
                "msg": {}
            }
        },
...
        "controller._ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "msg": {}
            }
        },
        "controller._ResponseExportArticle": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "content": {
                            "type": "string"
                        },
                        "filename": {
                            "type": "string"
                        },
                        "size": {
                            "type": "integer"
                        }
                    }
                },
                "msg": {}
            }
        },
        "controller._ResponseExportBatch": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "batch_id": {
                            "type": "string"
                        },
                        "file_count": {
                            "type": "integer"
                        },
                        "files": {},
                        "total_size": {
                            "type": "integer"
                        }
                    }
                },
                "msg": {}
            }
        },
...
        "controller._ResponseSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "data": {},
                "msg": {}
            }
        },
...
        "models.Comment": {
            "type": "object",
            "properties": {
                "article_id": {
                    "type": "string",
                    "example": "0"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "like_count": {
                    "type": "integer"
                },
                "parent_id": {
                    "type": "string",
                    "example": "0"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
...
        "models.ExportArticlesBatchRequest": {
            "type": "object",
            "required": [
                "article_ids"
            ],
            "properties": {
                "article_ids": {
                    "type": "array",
                    "maxItems": 50,
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "models.ParamCreateComment": {
            "type": "object",
            "required": [
                "article_id",
                "content"
            ],
            "properties": {
                "article_id": {
                    "type": "string",
                    "example": "0"
                },
                "content": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1
                },
                "parent_id": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
...
    }
}